+++ CHANGELOG.rst
@@ -1,7 +1,25 @@
 4.5.10 (unreleased)
 -------------------
 
+Fixes:
+
+- Fix to work with cockroach db >= 2.1
+  [vangheem]
+
 - Fix use of default factory for default value on content
+  [vangheem]
+
+
+New:
+
+- Provide warning when using insecure jwt secret in production
+  [vangheem]
+
+- Provide new `gen-key` for generating secure jwk keys
+  and jwt secrets
+  [vangheem]
+
+- Give warning when generating jwk key
   [vangheem]
 
 - Fix jwt implementation to use customized algorithm for encoding
@@ -12,6 +30,14 @@
 
 - Fix error message when trying to delete a concrete behavior
   [marcus29200]
+
+Changes:
+
+- Remove `utils.clear_conn_statement_cache`
+  [vangheem]
+
+- Lazy load dynamically generated jwk key
+  [vangheem]
 
 
 4.5.9 (2019-03-18)